// frontend/src/api/generated/models/inbound-slip-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipLineDetail } from './inbound-slip-line-detail';
// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipStatus } from './inbound-slip-status';
// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipType } from './inbound-slip-type';

/**
 * 
 * @export
 * @interface InboundSlipDetail
 */
export interface InboundSlipDetail {
    /**
     * 入荷伝票ID
     * @type {number}
     * @memberof InboundSlipDetail
     */
    'id': number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'slipNumber': string;
    /**
     * 
     * @type {InboundSlipType}
     * @memberof InboundSlipDetail
     */
    'slipType': InboundSlipType;
    /**
     * 振替伝票番号（倉庫振替の場合のみ）
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'transferSlipNumber'?: string | null;
    /**
     * 倉庫ID
     * @type {number}
     * @memberof InboundSlipDetail
     */
    'warehouseId': number;
    /**
     * 倉庫コード
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'warehouseCode': string;
    /**
     * 倉庫名
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'warehouseName': string;
    /**
     * 仕入先ID
     * @type {number}
     * @memberof InboundSlipDetail
     */
    'partnerId'?: number | null;
    /**
     * 仕入先コード
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'partnerCode'?: string | null;
    /**
     * 仕入先名
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'partnerName'?: string | null;
    /**
     * 入荷予定日
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'plannedDate': string;
    /**
     * 
     * @type {InboundSlipStatus}
     * @memberof InboundSlipDetail
     */
    'status': InboundSlipStatus;
    /**
     * 備考
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'note'?: string | null;
    /**
     * キャンセル日時
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'cancelledAt'?: string | null;
    /**
     * キャンセル者ID
     * @type {number}
     * @memberof InboundSlipDetail
     */
    'cancelledBy'?: number | null;
    /**
     * 登録日時
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'createdAt': string;
    /**
     * 登録者ID
     * @type {number}
     * @memberof InboundSlipDetail
     */
    'createdBy': number;
    /**
     * 登録者名
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'createdByName'?: string;
    /**
     * 更新日時
     * @type {string}
     * @memberof InboundSlipDetail
     */
    'updatedAt': string;
    /**
     * 更新者ID
     * @type {number}
     * @memberof InboundSlipDetail
     */
    'updatedBy': number;
    /**
     * 
     * @type {Array<InboundSlipLineDetail>}
     * @memberof InboundSlipDetail
     */
    'lines': Array<InboundSlipLineDetail>;
}