// frontend/src/api/generated/models/inbound-slip-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 入荷伝票ステータス
 * @export
 * @enum {string}
 */

export const InboundSlipStatus = {
    Planned: 'PLANNED',
    Confirmed: 'CONFIRMED',
    Inspecting: 'INSPECTING',
    PartialStored: 'PARTIAL_STORED',
    Stored: 'STORED',
    Cancelled: 'CANCELLED'
} as const;

export type InboundSlipStatus = typeof InboundSlipStatus[keyof typeof InboundSlipStatus];